nav: unify TabBar/SubTabBar visuals via shared TabStrip
Extract src/components/tabs/TabStrip.tsx as the single renderer for
both tab levels. Active entry: bold label + theme.primary fg on a
backgroundElement block; inactive: textMuted, no block. paddingX=2,
marginRight=1 throughout.

TabBar drops the leading digit prefix (idx() + borderSubtle span) and
the magic attributes={3}; hint still advertises <leader> N. SubTabBar
drops the bullet + theme.accent treatment and becomes a thin pass-
through to TabStrip. Callers and nav chords unchanged.

test/app.test.tsx: update the two assertions that matched on the
"1 Chat" prefix.

Spec: docs/nav.md § Tab Bars
kanban: herm t_ae78f566
